+++ src/InformaticsGateway/Services/Http/Startup.cs
@@ -14,16 +14,22 @@
  * limitations under the License.
  */
 
+using System.Linq;
+using System.Net.Mime;
 using System.Text.Json;
 using System.Text.Json.Serialization;
 using FellowOakDicom.Serialization;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Routing;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.OpenApi.Models;
+using Monai.Deploy.InformaticsGateway.Database;
+using Monai.Deploy.InformaticsGateway.Services.Fhir;
 
 namespace Monai.Deploy.InformaticsGateway.Services.Http
 {
@@ -42,7 +48,26 @@ namespace Monai.Deploy.InformaticsGateway.Services.Http
 #pragma warning restore CA1822 // Mark members as static
         {
             services.AddHttpContextAccessor();
-            services.AddControllers().AddJsonOptions(opts =>
+            services.AddControllers(opts =>
+            {
+                opts.RespectBrowserAcceptHeader = true;
+                var jsonSerializerOptions = new JsonSerializerOptions
+                {
+                    DefaultIgnoreCondition = JsonIgnoreCondition.WhenWritingNull,
+                    PropertyNameCaseInsensitive = true,
+                    DictionaryKeyPolicy = JsonNamingPolicy.CamelCase,
+                    PropertyNamingPolicy = JsonNamingPolicy.CamelCase,
+                    NumberHandling = JsonNumberHandling.AllowReadingFromString,
+                    WriteIndented = false
+                };
+
+                jsonSerializerOptions.Converters.Add(new JsonStringEnumMemberConverter(JsonNamingPolicy.CamelCase, false));
+                jsonSerializerOptions.Converters.Add(new DicomJsonConverter(writeTagsAsKeywords: false, autoValidate: false));
+                jsonSerializerOptions.Converters.Add(new JsonStringEnumConverter(JsonNamingPolicy.CamelCase, false));
+                opts.OutputFormatters.Add(new FhirJsonFormatters(jsonSerializerOptions));
+                opts.OutputFormatters.Add(new FhirXmlFormatters());
+            })
+            .AddJsonOptions(opts =>
             {
                 opts.JsonSerializerOptions.DefaultIgnoreCondition = JsonIgnoreCondition.WhenWritingNull;
                 opts.JsonSerializerOptions.PropertyNameCaseInsensitive = true;
@@ -53,6 +78,12 @@ namespace Monai.Deploy.InformaticsGateway.Services.Http
                 opts.JsonSerializerOptions.Converters.Add(new JsonStringEnumMemberConverter(JsonNamingPolicy.CamelCase, false));
                 opts.JsonSerializerOptions.Converters.Add(new DicomJsonConverter(writeTagsAsKeywords: false, autoValidate: false));
                 opts.JsonSerializerOptions.Converters.Add(new JsonStringEnumConverter(JsonNamingPolicy.CamelCase, false));
+            })
+            .AddXmlSerializerFormatters();
+
+            services.Configure<RouteOptions>(options =>
+            {
+                options.ConstraintMap.Add("fhirResource", typeof(FhirResourceTypesRouteConstraint));
             });
 
             services.AddSwaggerGen(c =>
@@ -74,6 +105,10 @@ namespace Monai.Deploy.InformaticsGateway.Services.Http
                     return result;
                 };
             });
+
+            services.AddHealthChecks()
+                .AddCheck<MonaiHealthCheck>("Informatics Gateway Services")
+                .AddDbContextCheck<InformaticsGatewayContext>("Database");
         }
 
 #pragma warning disable CA1822 // Mark members as static
@@ -89,6 +124,24 @@ namespace Monai.Deploy.InformaticsGateway.Services.Http
             }
 
             app.UseRouting();
+            app.UseHealthChecks("/health", new Microsoft.AspNetCore.Diagnostics.HealthChecks.HealthCheckOptions
+            {
+                ResponseWriter = async (context, report) =>
+                {
+                    var result = System.Text.Json.JsonSerializer.Serialize(new
+                    {
+                        status = report.Status.ToString(),
+                        checks = report.Entries.Select(c => new
+                        {
+                            check = c.Key,
+                            result = c.Value.Status.ToString()
+                        }),
+                    });
+
+                    context.Response.ContentType = MediaTypeNames.Application.Json;
+                    await context.Response.WriteAsync(result);
+                }
+            });
 
             app.UseEndpoints(endpoints =>
             {